Move all *Request objects to the corresponding *_server.go file
