+++ rpc/master/docker_server.go
@@ -13,6 +13,18 @@
 
 package master
 
+// UpgradeDockerRequest are options for upgrading/migrating the docker registry.
+type UpgradeDockerRequest struct {
+	Endpoint string
+	Override bool
+}
+
+// DockerOverrideRequest are options for replacing an image in the docker registry
+type DockerOverrideRequest struct {
+	OldImage string
+	NewImage string
+}
+
 // ResetRegistry pulls from the configured docker registry and updates the
 // index.
 func (s *Server) ResetRegistry(req struct{}, reply *int) error {
