+++ rpc/master/publicendpoint_server.go
@@ -17,6 +17,17 @@ import (
 	"github.com/control-center/serviced/domain/servicedefinition"
 )
 
+// Defines a request to add a port public endpoints to a service
+type PublicEndpointRequest struct {
+	Serviceid    string
+	EndpointName string
+	Name         string
+	UseTLS       bool
+	Protocol     string
+	IsEnabled    bool
+	Restart      bool
+}
+
 // Adds a port public endpoint to a service.
 func (s *Server) AddPublicEndpointPort(request *PublicEndpointRequest, reply *servicedefinition.Port) error {
 	port, err := s.f.AddPublicEndpointPort(s.context(), request.Serviceid, request.EndpointName, request.Name,
