+++ rpc/master/service_server.go
@@ -13,7 +13,31 @@
 
 package master
 
-import "github.com/control-center/serviced/domain/service"
+import (
+	"time"
+
+	"github.com/control-center/serviced/domain/service"
+)
+
+type ServiceUseRequest struct {
+	ServiceID   string
+	ImageID     string
+	ReplaceImgs []string
+	Registry    string
+	NoOp        bool
+}
+
+type WaitServiceRequest struct {
+	ServiceIDs []string
+	State      service.DesiredState
+	Timeout    time.Duration
+	Recursive  bool
+}
+
+type EvaluateServiceRequest struct {
+	ServiceID  string
+	InstanceID int
+}
 
 // Use a new image for a given service - this will pull the image and tag it
 func (s *Server) ServiceUse(request *ServiceUseRequest, response *string) error {
